Add namespaces for all indri classes. Add Trevor's 1.1 concurrency changes. Fix cache handling in BulkTreeReader. Add weighted or node and weighted extents.
git-svn-id: svn+ssh://svn.code.sf.net/p/lemur/code/indri/trunk@384 14020d50-3050-45fe-8e0c-5ed7005ca764

+++ contrib/lemur/include/lemur/RVLCompress.hpp
@@ -126,10 +126,10 @@ inline const char* RVLCompress::decompress_longlong( const char* source, INT64& 
 
   for( i=0; i<10; i++ ) {
     if( source[i] & terminator ) {
-      data |= ((source[i] & mask) << 7*i);
+      data |= (INT64(source[i] & mask) << 7*i);
       break;
     } else {
-      data |= (source[i] << 7*i);
+      data |= (INT64(source[i]) << 7*i);
     }
   }
